Move even more widgets to ui lib
