+++ src/ui/misc/noshortcutaction.cpp
