+++ src/ui/widgets/canvassettingswidget.cpp
