+++ src/ui/widgets/performancesettingswidget.h
@@ -1,7 +1,34 @@
+/*
+#
+# Friction - https://friction.graphics
+#
+# Copyright (c) Friction contributors
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+# See 'README.md' for more information.
+#
+*/
+
+// Fork of enve - Copyright (C) 2016-2020 Maurycy Liebner
+
 #ifndef PERFORMANCESETTINGSWIDGET_H
 #define PERFORMANCESETTINGSWIDGET_H
 
-#include "settingswidget.h"
+#include "ui_global.h"
+
+#include "widgets/settingswidget.h"
 #include <QSpinBox>
 #include <QLabel>
 #include <QCheckBox>
@@ -9,7 +36,7 @@
 #include <QComboBox>
 #include <QList>
 
-class PerformanceSettingsWidget : public SettingsWidget
+class UI_EXPORT PerformanceSettingsWidget : public SettingsWidget
 {
 public:
     explicit PerformanceSettingsWidget(QWidget *parent = nullptr);
